project_name/requirements/local.txt
# LOCAL dependencies go here.
-r test.txt

django-debug-toolbar==1.2.1
